refactor(server): decompose api-routes into route handler files
Splits the 584-LOC api-routes.ts monolith into 12 focused files under
src/server/mcp/routes/. Each route exports a make*Handler() factory;
shared utilities (sendApiError, errorCodeToHttpStatus, validators) live
in routes/_shared.ts. api-routes.ts becomes a thin orchestrator (~100
LOC) that wires factories and re-exports symbols for existing consumers.
Zero behavioral changes; all test imports continue to work unchanged.
