+++ src/server/mcp/routes/apply-changes.ts
@@ -1,0 +1,33 @@
+import type { Request, Response } from "express";
+import type { Handler } from "../api-routes.js";
+import { applyChangesCore } from "../docx-apply.js";
+import { sendApiError } from "./_shared.js";
+
+export function makeApplyChangesHandler(): Handler {
+  return async (req: Request, res: Response) => {
+    const { documentId, author, backupPath } = (req.body ?? {}) as Record<string, unknown>;
+    if (documentId !== undefined && typeof documentId !== "string") {
+      res.status(400).json({ error: "BAD_REQUEST", message: "documentId must be a string" });
+      return;
+    }
+    if (author !== undefined && typeof author !== "string") {
+      res.status(400).json({ error: "BAD_REQUEST", message: "author must be a string" });
+      return;
+    }
+    if (backupPath !== undefined && typeof backupPath !== "string") {
+      res.status(400).json({ error: "BAD_REQUEST", message: "backupPath must be a string" });
+      return;
+    }
+
+    try {
+      const result = await applyChangesCore(
+        documentId as string | undefined,
+        author as string | undefined,
+        backupPath as string | undefined,
+      );
+      res.json({ data: result });
+    } catch (err: unknown) {
+      sendApiError(res, err);
+    }
+  };
+}
