+++ src/server/mcp/routes/open.ts
@@ -1,0 +1,20 @@
+import type { Request, Response } from "express";
+import type { Handler } from "../api-routes.js";
+import { openFileByPath } from "../file-opener.js";
+import { sendApiError } from "./_shared.js";
+
+export function makeOpenHandler(): Handler {
+  return async (req: Request, res: Response) => {
+    const { filePath, force } = (req.body ?? {}) as Record<string, unknown>;
+    if (!filePath || typeof filePath !== "string") {
+      res.status(400).json({ error: "BAD_REQUEST", message: "filePath is required" });
+      return;
+    }
+    try {
+      const result = await openFileByPath(filePath, { force: force === true });
+      res.json({ data: result });
+    } catch (err: unknown) {
+      sendApiError(res, err);
+    }
+  };
+}
